Do not inspect noise/component via mouse click on graph by default
This fixes an issue (or rather works around it) with graph losing
focus on changing node connections. This used to work fine
but something has changed within the engine so existing behavior
doesn't work reliably anymore.

The noise properties can be edited by pressing the first button in edited
path now, while the components can be just selected in order for them
to be saved externally, for instance.